+++ prov/slurm/executor.go
@@ -94,16 +94,16 @@ func (e *defaultExecutor) ExecDelegate(ctx context.Context, cfg config.Configura
 	operation := strings.ToLower(delegateOperation)
 	switch {
 	case operation == "install":
-		err = e.installNode(ctx, locationProps, deploymentID, nodeName, instances, operation)
+		err = e.installNode(ctx, cfg, locationProps, deploymentID, nodeName, instances, operation)
 	case operation == "uninstall":
-		err = e.uninstallNode(ctx, locationProps, deploymentID, nodeName, instances, operation)
+		err = e.uninstallNode(ctx, cfg, locationProps, deploymentID, nodeName, instances, operation)
 	default:
 		return errors.Errorf("Unsupported operation %q", delegateOperation)
 	}
 	return err
 }
 
-func (e *defaultExecutor) installNode(ctx context.Context, locationProps config.DynamicMap, deploymentID, nodeName string, instances []string, operation string) error {
+func (e *defaultExecutor) installNode(ctx context.Context, cfg config.Configuration, locationProps config.DynamicMap, deploymentID, nodeName string, instances []string, operation string) error {
 	for _, instance := range instances {
 		err := deployments.SetInstanceStateWithContextualLogs(events.AddLogOptionalFields(ctx, events.LogOptionalFields{events.InstanceID: instance}), deploymentID, nodeName, instance, tosca.NodeStateCreating)
 		if err != nil {
@@ -115,10 +115,10 @@ func (e *defaultExecutor) installNode(ctx context.Context, locationProps config.
 		return err
 	}
 
-	return e.createInfrastructure(ctx, locationProps, deploymentID, nodeName, infra)
+	return e.createInfrastructure(ctx, cfg, locationProps, deploymentID, nodeName, infra)
 }
 
-func (e *defaultExecutor) uninstallNode(ctx context.Context, locationProps config.DynamicMap, deploymentID, nodeName string, instances []string, operation string) error {
+func (e *defaultExecutor) uninstallNode(ctx context.Context, cfg config.Configuration, locationProps config.DynamicMap, deploymentID, nodeName string, instances []string, operation string) error {
 	for _, instance := range instances {
 		err := deployments.SetInstanceStateWithContextualLogs(events.AddLogOptionalFields(ctx, events.LogOptionalFields{events.InstanceID: instance}), deploymentID, nodeName, instance, tosca.NodeStateDeleting)
 		if err != nil {
@@ -130,16 +130,16 @@ func (e *defaultExecutor) uninstallNode(ctx context.Context, locationProps confi
 		return err
 	}
 
-	return e.destroyInfrastructure(ctx, locationProps, deploymentID, nodeName, infra)
+	return e.destroyInfrastructure(ctx, cfg, locationProps, deploymentID, nodeName, infra)
 }
 
-func (e *defaultExecutor) createInfrastructure(ctx context.Context, locationProps config.DynamicMap, deploymentID, nodeName string, infra *infrastructure) error {
+func (e *defaultExecutor) createInfrastructure(ctx context.Context, cfg config.Configuration, locationProps config.DynamicMap, deploymentID, nodeName string, infra *infrastructure) error {
 	events.WithContextOptionalFields(ctx).NewLogEntry(events.LogLevelINFO, deploymentID).RegisterAsString("Creating the slurm infrastructure")
 	var g errgroup.Group
 	for _, compute := range infra.nodes {
 		func(ctx context.Context, comp *nodeAllocation) {
 			g.Go(func() error {
-				return e.createNodeAllocation(ctx, locationProps, comp, deploymentID, nodeName)
+				return e.createNodeAllocation(ctx, cfg, locationProps, comp, deploymentID, nodeName)
 			})
 		}(events.AddLogOptionalFields(ctx, events.LogOptionalFields{events.InstanceID: compute.instanceName}), compute)
 	}
@@ -155,13 +155,13 @@ func (e *defaultExecutor) createInfrastructure(ctx context.Context, locationProp
 	return nil
 }
 
-func (e *defaultExecutor) destroyInfrastructure(ctx context.Context, locationProps config.DynamicMap, deploymentID, nodeName string, infra *infrastructure) error {
+func (e *defaultExecutor) destroyInfrastructure(ctx context.Context, cfg config.Configuration, locationProps config.DynamicMap, deploymentID, nodeName string, infra *infrastructure) error {
 	events.WithContextOptionalFields(ctx).NewLogEntry(events.LogLevelINFO, deploymentID).RegisterAsString("Destroying the slurm infrastructure")
 	var g errgroup.Group
 	for _, compute := range infra.nodes {
 		func(ctx context.Context, comp *nodeAllocation) {
 			g.Go(func() error {
-				return e.destroyNodeAllocation(ctx, locationProps, comp, deploymentID, nodeName)
+				return e.destroyNodeAllocation(ctx, cfg, locationProps, comp, deploymentID, nodeName)
 			})
 		}(events.AddLogOptionalFields(ctx, events.LogOptionalFields{events.InstanceID: compute.instanceName}), compute)
 	}
@@ -177,12 +177,12 @@ func (e *defaultExecutor) destroyInfrastructure(ctx context.Context, locationPro
 	return nil
 }
 
-func (e *defaultExecutor) createNodeAllocation(ctx context.Context, locationProps config.DynamicMap, nodeAlloc *nodeAllocation, deploymentID, nodeName string) error {
+func (e *defaultExecutor) createNodeAllocation(ctx context.Context, cfg config.Configuration, locationProps config.DynamicMap, nodeAlloc *nodeAllocation, deploymentID, nodeName string) error {
 	events.WithContextOptionalFields(ctx).NewLogEntry(events.LogLevelINFO, deploymentID).RegisterAsString(fmt.Sprintf("Creating node allocation for: deploymentID:%q, node name:%q", deploymentID, nodeName))
 
 	// Return an sshClient configured using the user credentials provided in the yorc.nodes.slurm.Compute node definition,
 	// or if not provided, the user credentials specified in the Yorc configuration
-	sshClient, err := getSSHClient(nodeAlloc.credentials, locationProps)
+	sshClient, err := getSSHClient(cfg, nodeAlloc.credentials, locationProps)
 	if err != nil {
 		events.WithContextOptionalFields(ctx).NewLogEntry(events.LogLevelERROR, deploymentID).RegisterAsString(err.Error())
 		return err
@@ -335,12 +335,12 @@ func (e *defaultExecutor) createNodeAllocation(ctx context.Context, locationProp
 	return nil
 }
 
-func (e *defaultExecutor) destroyNodeAllocation(ctx context.Context, locationProps config.DynamicMap, nodeAlloc *nodeAllocation, deploymentID, nodeName string) error {
+func (e *defaultExecutor) destroyNodeAllocation(ctx context.Context, cfg config.Configuration, locationProps config.DynamicMap, nodeAlloc *nodeAllocation, deploymentID, nodeName string) error {
 	events.WithContextOptionalFields(ctx).NewLogEntry(events.LogLevelINFO, deploymentID).RegisterAsString(fmt.Sprintf("Destroying node allocation for: deploymentID:%q, node name:%q, instance name:%q", deploymentID, nodeName, nodeAlloc.instanceName))
 
 	// Return an sshClient configured using the user credentials provided in the yorc.nodes.slurm.Compute node definition,
 	// or if not provided, the user credentials specified in the Yorc configuration
-	sshClient, err := getSSHClient(nodeAlloc.credentials, locationProps)
+	sshClient, err := getSSHClient(cfg, nodeAlloc.credentials, locationProps)
 	if err != nil {
 		events.WithContextOptionalFields(ctx).NewLogEntry(events.LogLevelERROR, deploymentID).RegisterAsString(err.Error())
 		return err